Granger fully integrated first pass
